ci: refactor Filebeat service integration test.
Moved the BackgroundProcess class from within FilebeatService
and created a helpers module so that it can be re-used.
